+++ src/main/java/com/coolerpromc/resourcestrees/datagen/ModItemTagProvider.java
@@ -1,0 +1,28 @@
+package com.coolerpromc.resourcestrees.datagen;
+
+import com.coolerpromc.resourcestrees.block.ModBlocks;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.data.PackOutput;
+import net.minecraft.data.tags.ItemTagsProvider;
+import net.minecraft.data.tags.TagsProvider;
+import net.minecraft.tags.ItemTags;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.level.block.Block;
+import net.neoforged.neoforge.registries.DeferredBlock;
+
+import java.util.concurrent.CompletableFuture;
+
+public class ModItemTagProvider extends ItemTagsProvider {
+    public ModItemTagProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> lookupProvider, CompletableFuture<TagsProvider.TagLookup<Block>> blockTags) {
+        super(output, lookupProvider, blockTags);
+    }
+
+    @Override
+    protected void addTags(HolderLookup.Provider provider) {
+        IntrinsicTagAppender<Item> builder = this.tag(ItemTags.SAPLINGS);
+
+        for (DeferredBlock<? extends Block> block : ModBlocks.SAPLINGS){
+            builder.add(block.asItem());
+        }
+    }
+}